refactor(core/config_models): step 2+3 - split in 3 moduli专项ati
Crea due nuovi moduli专项ati:

shared_models.py (~210 righe):
- 8 model Pydantic (TimeCoverage, DatasetBlock, SupportDatasetConfig, ...)
- ConfigDeprecation dataclass + _CONFIG_DEPRECATIONS
- parse_bool, ensure_str_list (coercion)
- _SAFE_SQL_IDENTIFIER_RE

policy.py (~130 righe):
- _TOP_LEVEL_ALLOWED_KEYS
- _declared_model_keys
- _normalize_legacy_payload
- _emit_deprecation_notice, _emit_unknown_keys_notice
- _warn_or_reject_unknown_keys

common.py ora thin facade (~95 righe):
- re-export da sub-moduli per backward compat
- consumer esistenti importano da common, non da sub-moduli

Risultato netto: path_normalization + shared_models + policy = 3 moduli专项ati
oltre a _loader.py. common.py come thin facade. Nessun contract cambiato.

63/63 test passati, ruff all green.
